+++ gemtext_to_html/Makefile
@@ -11,6 +11,7 @@ test: text_test
 
 build/gemtext_to_html_test: src/gemtext_to_html_test.c
 build/gemtext_to_html_test: src/gemtext_to_html.h
+build/gemtext_to_html_test: src/gemtext_to_html.c
 build/gemtext_to_html_test: src/arena.c
 build/gemtext_to_html_test: src/arena.h
 build/gemtext_to_html_test: src/astring.c
@@ -18,14 +19,20 @@ build/gemtext_to_html_test: src/astring.h
 build/gemtext_to_html_test: src/testing.h
 build/gemtext_to_html_test: src/text.h
 build/gemtext_to_html_test: | build
-	cc -g -Og -o "$@" src/gemtext_to_html_test.c src/arena.c src/astring.c
+	cc -g -Og -o "$@" \
+		src/gemtext_to_html_test.c \
+		src/gemtext_to_html.c \
+		src/arena.c \
+		src/astring.c
 
 build/text_test: src/text_test.c
 build/text_test: src/text.h
 build/text_test: src/arena.c
 build/text_test: src/arena.h
 build/text_test: | build
-	cc -g -Og -o "$@" src/text_test.c src/arena.c
+	cc -g -Og -o "$@" \
+		src/text_test.c \
+		src/arena.c
 
 .PHONY: mem_check
 mem_check:
